add menu item delete req, move code to separate class
